Исправили недочеты в контроллерах: убрали лишние логи. Перенесли обработку ошибок из классов бинов в класс глобального обработчика ошибок. Добавили обработку исключений в класс глобального обработчика. Покрыли юнит-тестами TransactionService. Еще некоторые минорные изменение/недочеты в других классах.

+++ app/src/main/java/com/bankingsystem/app/model/ExchangeRateDTO.java
@@ -7,5 +7,5 @@ import java.math.BigDecimal;
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ExchangeRateDTO {
-    private double rate;
+    private BigDecimal rate;
 }
